requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
blas=1.0=mkl
bottleneck=1.3.2=py39h7cc1a96_1
ca-certificates=2021.7.5=haa95532_1
certifi=2021.5.30=py39haa95532_0
colorama=0.4.4=pypi_0
intel-openmp=2021.3.0=haa95532_3372
lxml=4.8.0=pypi_0
mkl=2021.3.0=haa95532_524
mkl-service=2.4.0=py39h2bbff1b_0
mkl_fft=1.3.0=py39h277e83a_2
mkl_random=1.2.2=py39hf11a4ad_0
mpmath=1.2.1=py39haa95532_0
numexpr=2.7.3=py39hb80d3ca_1
numpy=1.20.3=py39ha4e8547_0
numpy-base=1.20.3=py39hc2deb75_0
openssl=1.1.1k=h2bbff1b_0
pandas=1.3.0=py39hd77b12b_0
pillow=9.0.0=pypi_0
pip=21.1.3=py39haa95532_0
pypdf2=1.26.0=pypi_0
python=3.9.6=h6244533_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-docx=0.8.11=pypi_0
pytz=2021.1=pyhd3eb1b0_0
setuptools=52.0.0=py39haa95532_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.36.0=h2bbff1b_0
sympy=1.8=py39haa95532_0
tqdm=4.62.3=pypi_0
tzdata=2021a=h52ac0ba_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.36.2=pyhd3eb1b0_0
wincertstore=0.2=py39h2bbff1b_0
